+++ inc/customizer-selective-refresh.php
@@ -292,6 +292,11 @@ function onepress_customizer_partials( $wp_customize ) {
         'onepress_primary_color',
         'onepress_menu_item_padding',
 
+        'onepress_page_cover_color',
+        'onepress_page_cover_overlay',
+        'onepress_page_cover_pd_top',
+        'onepress_page_cover_pd_bottom',
+
         'onepress_header_bg_color',
         'onepress_menu_color',
         'onepress_menu_hover_color',