+++ changelog.md
@@ -21,6 +21,7 @@ CHANGELOG
 * NEW: Footer copyright link colors.
 * NEW: Hero rotating text and background color for layout 1.
 * NEW: Add an option to disable default google theme in theme.
+* NEW: Add page header cover.
 * IMPROVED: Latest bootstrap - v4.0.0-alpha.6.
 * IMPROVED: Carousel and RTL.
 * IMPROVED: Change Actions Required text to Recommended Actions.